+++ src/main/java/dev/gacbl/logicore/data/ModBlockTagProvider.java
@@ -9,6 +9,7 @@ import dev.gacbl.logicore.blocks.datacable.DataCableModule;
 import dev.gacbl.logicore.blocks.datacenter.DatacenterModule;
 import dev.gacbl.logicore.blocks.datacenter_port.DatacenterPortModule;
 import dev.gacbl.logicore.blocks.generator.GeneratorModule;
+import dev.gacbl.logicore.blocks.repair_station.RepairStationModule;
 import dev.gacbl.logicore.blocks.research_station.ResearchStationModule;
 import dev.gacbl.logicore.blocks.serverrack.ServerRackModule;
 import dev.gacbl.logicore.core.ModTags;
@@ -116,6 +117,14 @@ public class ModBlockTagProvider extends BlockTagsProvider {
         tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "swapper_deny"))).add(ResearchStationModule.RESEARCH_STATION.get());
         tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("c", "relocation_not_supported"))).add(ResearchStationModule.RESEARCH_STATION.get());
         tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("mekanism", "cardboard_blacklist"))).add(ResearchStationModule.RESEARCH_STATION.get());
+
+        tag(BlockTags.MINEABLE_WITH_PICKAXE).add(RepairStationModule.REPAIR_STATION.get());
+        tag(BlockTags.NEEDS_IRON_TOOL).add(RepairStationModule.REPAIR_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "eclipsegate_deny"))).add(RepairStationModule.REPAIR_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "phase_deny"))).add(RepairStationModule.REPAIR_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "swapper_deny"))).add(RepairStationModule.REPAIR_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("c", "relocation_not_supported"))).add(RepairStationModule.REPAIR_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("mekanism", "cardboard_blacklist"))).add(RepairStationModule.REPAIR_STATION.get());
 
         setupOptionalTags();